Refactored model into three separate models, for RPN, Faster RCNN and Mask-RCNN. Partial Detectron compatibility; include centred anchor boxes and RoIAlign. Supports batch sizes larger than 1 for training and inference. Requires PyTorch 0.4.

+++ maskrcnn/nms/src/cuda/nms_kernel.h
@@ -6,10 +6,10 @@ extern "C" {
 #endif
 
 #define DIVUP(m,n) ((m) / (n) + ((m) % (n) > 0))
-int const threadsPerBlock = sizeof(unsigned long long) * 8;
+int const threadsPerBlock = sizeof(unsigned long int) * 8;
 
 void _nms(int boxes_num, float * boxes_dev,
-          unsigned long long * mask_dev, float nms_overlap_thresh);
+          long int * mask_dev, float nms_overlap_thresh);
 
 #ifdef __cplusplus
 }
